Move files depending the type and create a gitignore
