--- localMatchApp/logIn.cpp
@@ -1,31 +1,0 @@
-#include "logIn.h"
-
-using namespace std;
-
-logIn::logIn()
-{
-    //Connect to database.
-}
-
-logIn::~logIn()
-{
-    //Disconnect to database.
-}
-
-int logIn::signUp(const string i_user, const string i_password)
-{
-  //Send information to database.
-  return 0;
-}
-
-bool logIn::isEqualPass(const string i_pass, const string i_rePass)
-{
-  //Compare strings and return TRUE for equal.
-  return true;
-}
-
-int logIn::entry(const string i_user, const string i_password)
-{
-  //Search the user and compare password
-  return 0;
-}
